parted.make - support versions 3.1 and 3.4
This adds conditional logic so that:
 armv8a compiles with parted 3.4
 x86_64 compiles with parted 3.1

And adds a parted-configure-help option to see parted
configuration options.

The reason for the two version is that armv8a, which uses
glibc for its C libraries, could build 3.4, not 3.1, and
x86_64, which uses uClibc for it's C libraries, does not
support certain 'locale' options in any way that could be
used by the 3.4 version of parted ( look for compile errors
with 'atari.c' )

While I would like to see both builds using the same version,
after putting a few days of effort in to re-configuring
crosstool-ng to reconfigure uClibc (and changing the uClibc
configuration file), and examining parted configuration options,
and reading comments about how support for this was mostly
workarounds, I concluded that uClibc just wasn't ready for this,
and kept the older version for x86_64.

uClibc configuration values of interest (either set or not) are
UCLIBC_HAS_LOCALE
UCLIBC_HAS_XLOCALE
UCLIBC_BUILD_ALL_LOCALE
UCLIBC_BUILD_MINIMAL_LOCALE

crosstool-ng configuration values of interest are:
CT_LIBC_UCLIBC_LOCALES

Some permutation of the above may work, and I'd love to
see a pull request that fixes this.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>