kexec-tools - upgrade from 2.0.9 to 2.0.22
This upgrade supports armv8a, x86 and is part of the upgrade to
a 5.4 kernel, GCC 8.3, etc.

Note that the patches have been disabled here.
One was for PowerPC, which may no longer be relevant
(And there is certainly no way to test it at the moment, so it
being a provable liability counts for more than it being a possible
asset)
And the other was for arm64 support, from 2013. Given that it
builds and runs fine for arm64, that patch has been disabled.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>