emulation README - update for armv8a targets
This updates the README-onie-vm.md file with instructions
for running the qemu_armv8a build in emulation using
the onie-vm.sh script.

It also changed a few of the X86 instructions to be X86 specific,
now that there is more than one possible emulation target.

onie-build-targets.json has been updated.

Testing
--------
cat onie-build-targets.json | jq '.'

Builds, runs with onie-vm.sh, and demo OS installs
 make MACHINE=qemu_armv8a -j4 all demo
 onie-vm.sh run --machine-name qemu_armv8a --m-bios-uefi --m-embed-onie
  Once installed and in ONIE rescue mode, onie-nos-install installed the
  ARM64 Demo OS, and it booted to a prompt

Making sure nothing broke for KVM
 make MACHINE=kvm_x86_64 -j4 all demo
 ./onie-vm.sh run --m-usb-drive --m-bios-uefi --m-secure --m-hd-clean --m-bios-clean --m-onie-iso ../build/images/onie-recovery-x86_64-kvm_x86_64-r0.iso
 In rescue mode, installed Demo OS for x86
 Booted to EFI shell and installed keys
 Enabled Secure Boot in menu
 Rebooted
 ONIE asked for password,
   mokutil --sb-status showed Secure Boot active.

Final notes
-----------------
I have no doubt there is plenty of room for optimization,
improvement, and bugs to be fixed here, but I also don't
want the perfect to be the enemy of the good, and at least
with functional code available, others will be more able
to improve upon it.

I'd like to add a special thanks to
Shi Lei <shil@centecnetworks.com>, for PR #956, which
got me _something_ that built along with a valid
UEFI BIOS that got me on the right track to fixing this.

As of this commit qemu_armv8a is fixed.

Closes #940

Signed-off-by: Alex Doyle <adoyle@nvidia.com>