Makefile -add debian-prepare-build-host
The debian-prepare-build-host target may need to be
depreciated, or perhaps restricted to only configuring
for the latest release of Debian, as package names
and dependencies change release to release.

This adds a suggestion about using Docker contianers for
build (where a build environment's release version can be assured)
and a comment about reading the FAQ before building the
target.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>