ARM64 emulation - use KVM if emulating on an ARM64  system.
This uses kvm and the host processor to run the emulation if
ONIE is executing on an ARM64 system, rather than emulating a
processor.

Users will have to be members of the host system's kvm group,
for this to work, but the speed up is _quite_ noticeable.

Note that ONIE won't build on an ARM64 system, but the emulation
environment can be copied over and run there.

Testing:
--------
Worked on a multi core Amazon Web Serivces ARM64 system running
Ubuntu and a Raspberry Pi 4 running Debian 11 Bullseye.

Signed-off-by: Alex Doyle <aodyle@nvidia.com>