Emulation - add 'export-emulation' option to onie-vm
This option creates a tar file (with optional custom name)
that contains the ./emulation and ./build/images directories.
Those are everything ONIE emulation needs to run.
While this feature may seem trivial, (and it is), it
becomes extremely useful when one has to use it more than
a few times a day...

To run, just extract and cd into the ./emulation directory
and run the onie-vm.sh script.

Uses for this include:
 - Providing a run time environment for other developers or
    testers who may be waiting for access to hardware, but
	do not care about the rest of ONIE's build environment.

 - Faster emulation by using a different host processor.
     For example, ARM64 emulation speeds up quite a bit if it
     can use KVM and the host processor, rather than emulate
     a CPU, so copying the emulation files can be a big
	 time server.
	 Testing this, a Raspberry PI 4 had comparable performance
	 to a high end X86 laptop running ARM64 emulation, and
	 a multi core AWS ARM64 instance was MUCH faster.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>