Signing keys - add 'signing-keys-generate' as PHONY make target
This adds 'signing-keys-generate' to the list of PHONY signing
keys targets. This target does the acutal key generation, and
is used by 'signing-keys-install', which takes the generated keys
and adds them to the virtual USB drive for demonstration in emulation.

However, it may be that one just wants the keys and not the last
step ( expecially because the signing-keys-install requires a
DUE container to be run with
 due --run --dockerarg privileged
To be able to loopback mount the virtual USB device where the keys
would be stored.)

Testing:
--------
Generated keys with:
  make MACHINE=kvm_x86_64 signing-keys-generate

Signed-off-by: Alex Doyle <adoyle@nvidia.com>