Emulation - allow x86 to run without kvm
Update the x86 runtime options to look for a
loaded kvm kernel module. If it is not present,
either KVM is not installed, or the script is
already executing in a virtualization environment,
so do NOT specify the kvm and host flags to QEMU.

Discovered this recording a demo on a fresh Debian
install on a VirtualBox instance.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>