GitHub docs: add _static/language_data.js
The 1.8.4-1 version of sphinx that comes with Debian 10
generates this file when building docs.
As a Debain 10 build environment is now standard, this
commit addresses the problem of a 'make publish' failing
because of an unrecognized file.

From here on out, documentation should be built from a Debian 10
environment.

Testing:
 build/docs/html copied to a web server directory.
 Was browseable and search worked.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>