Makefile - set PYTHON environment variable
For build environments prior to Debian 11, /usr/bin/python was a
default for python 2x. In Debian 11, Python presents as:
 /usr/bin/python2
 /usr/bin/python3

This causes a configuration failure for the GRUB build.

With this fix the top level Makefile has been updated to
export PYTHON=python2 if /usr/bin/python is not present,
assuming that any program looking for the old implementation
of Python will want python2

Testing
-------
Built the kvm_x86_64 and qemu_armv8a virtual targets
in both Debian10 and Debian11 DUE containers.

Both builds ran in emulation.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>