onie/emulation - support third party virtual machines with --m-onie-arch
This addresses the case of a developer creating their own
virtual machine target in addition to kvm_x86_64 and qemu_armv8a

The emulation script set a number of variables for architecture
specific emulation based off the machine types. This commit
changes the defaults to be set by archiecture type with --m-onie-arch

This way specifiying a machine name and architecture can be used
to run a new virtual target.
kvm_x86_64 and qemu_armv8a will still be recognized, and their
default values for emulation will be set without having to
speify the processor architecture.

Example invocation:
 Embed ONIE on new arm64 virtual machine 'mynewmachine'

 ./onie-vm.sh run --m-bios-uefi --m-bios-clean --m-hd-clean \
   --m-embed-onie --machine-name mynewmachine --m-onie-arch arm64

Testing
-------
 Used a private machine build that was a copy/paste/edit of
 onie/machine/qemu_armv8a to fufull the 'mynewmachine' requirement.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>