Add unzip to non-DUE build environments
This adds unzip to the Dockerfile that creates the
original Debian 9 Docker build environment, and
to the make target for configuring a host system.

In testing, the Debian 9 builds from the 2021.08
branch / Debian9BuildEnvironment tag seemed to build
fine with:
 make -j4 MACHINE=kvm_x86_64 all recovery-iso

However the Debian 10/11 DUE containers have unzip present,
so this is clearly needed for some builds.

In any event, adding this won't hurt anything, and, as
it has the potential to reduce developer frustration,
that's a clear benefit.

fix: add unzip to contrib/build-env/dockerfile
fix: add unzip to Makefile's debian-prepare-build-host target

Signed-off-by: Chris Cummings <chris@cummings.tech>
Signed-off-by: Alex Doyle <alexddoyle@gmail.com>