onie-build-targets.json - clean up formatting.
This removes some trailing whitespace and puts two
spaces after a "Debian9", entry to keep alignment if/when
that entry changes to "Debian10", with one space after.
This will keep focus on the line changed, rather than the
manufacturer paragraph it was in.

While it still looks messy, the policy for column alignment
is to apply it on a per-manufacturer basis, figuring that any
user lookups being done are going to be for a model from a
known manufacturer.

I'd also rather not have to readjust all entries if a switch
with a long name comes in, or introduce screen wrap by having
arbitrarily padded entries.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>