image.make: Fix ssh login by adding libnss_files.so to ARM64 sysroot image
This library is required for ssh localhost login to work, otherwise
local users are not recognized.

libnss_files.so is part of glibc NSS implementation, so it is added
to the SYSROOT_LIBS variable for all cases using glibc.
Currently, x86 builds do not use glibc, so this does not affect them.

For more details on NSS see:
 https://en.wikipedia.org/wiki/Name_Service_Switch.
This particular shared object provides access to plain text database files
like /etc/passwd and /etc/hosts.

On ARM64, DropBear failed to identify local users since the only way to
access /etc/passwd on glibc based systems is via libnss_files.so.
On X86 the login worked since uclibc does not provide NSS implementation,
as can be seen in
https://uclibc.org/downloads/Glibc_vs_uClibc_Differences.txt

Signed-off-by: Felix Radensky <fradensky@nvidia.com>
Signed-off-by: Alex Doyle <alexddoyle@gmail.com>