onie-build-targets.json - list 2022.02br compatible platforms.
This updates the list of platforms that build with the
(current) master branch that is entering lockout for the 2022.02
release candidate.

Note that this lists only the platforms that have been tested
recenlty with the master branch, so other platorms may build with
newer releases - either the 2021.11br component upgrade branch
that became the current master, or the older 2021.08br legacy
compatibility branch.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>