Add support for crosstool-NG 1.28.0
Add crosstool-NG 1.28.0 and update kvm_x86_64 to use it. For now the
changes are as minimal as possible so the GCC version is bumped from
8.3.0 to 8.5.0 and most other packages are held as they are. The various
debugging tools have been disabled as these are best built separately
(there are also some complications with GDB and python versions).

Subsequent changes should be possible to make use of the newer versions
of GCC, binutils, uClibc-ng and other packages.

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>